Change submit activity form to use WTForms
Soon realized as I started the edit option, that using WTForms would be
a better option in the long run to prevent some template duplication and
to also help with form validation.

Also re-styled the flash messages for info and errors.

Added consts.py file to store any constant variables to make them very easy
to find if they need to be changed.